Update scale filter and add workflow to settings
Add Jeff's slow_scale_filtering code to DSx2.
DSx2 will look for Jeff's slow_scale_filtering plugin and if it is not found it will load similar code from with in DSx2. Or if the plugin is found and not active, it will activate it.
This code smooths the weight curves. I have plans to use weight data for some feedback function at a later date, for things that will give feedback based on your shots, and I need clean data to do it. Plus, smooth curves are more pleasing to the eye and easier to read.

DSx2 workflow will be saved to ::settings(DSx2_workflow) so it will be saved to shot files